new stream for collective bargaining
