+++ backend/background/sources/collective_bargaining.py
@@ -1,8 +1,8 @@
 import asyncio
+import re
 from dotenv import load_dotenv
-from typing import List, AsyncIterator, Optional, Sequence, cast
-from bs4 import BeautifulSoup, Tag
-from playwright.async_api import Page, TimeoutError as PlaywrightTimeoutError
+from typing import List, AsyncIterator
+from playwright.async_api import Page
 from background.sources.shared import user_agent, get_browser_page
 
 from background.logger import setup_logger
@@ -13,196 +13,267 @@ from background.sources.document_stream import DocumentStream
 load_dotenv()  # Loads environment variables from .env
 logger = setup_logger()
 
-site_url: str = "https://ucnet.universityofcalifornia.edu"
-base_url: str = "https://ucnet.universityofcalifornia.edu/resources/employment-policies-contracts/bargaining-units/"
-
-# Helper class representing union details.
-
-
-class UnionDetail:
-    def __init__(self, name: str, code: str, url: str, scope: str) -> None:
-        self.name: str = name
-        self.code: str = code
-        self.url: str = url
-        # If URL is relative, prepend the site URL.
-        if not url.startswith("http"):
-            self.url = f"{site_url}{url}"
-        self.scope: str = scope  # Either "ucop" or the name of the campus
-
-    def __repr__(self) -> str:
-        return f"UnionDetail(name={self.name}, code={self.code}, url={self.url}, scope={self.scope})"
+BARGAINING_UNITS_URL = (
+    "https://ucnet.universityofcalifornia.edu/resources/employment-"
+    "policies-contracts/bargaining-units/"
+)
 
 
-class UcopCollectiveBargainingStream(DocumentStream):
+class UcnetCollectiveBargainingStream(DocumentStream):
     def __init__(self, source: Source) -> None:
         super().__init__(source)
         self.max_retries: int = 3
 
-    async def get_local_unions(self, url: str, page: Page) -> List[UnionDetail]:
-        """Retrieve local union details from the provided URL."""
-        await page.goto(url)
+    async def extract_pdf_links(self, page: Page) -> List[str]:
+        """Return a list of PDF URLs from the current page."""
+        pdf_links: List[str] = []
         try:
-            await page.wait_for_selector("#local", timeout=10000)
-        except PlaywrightTimeoutError as e:
-            logger.error(f"Error waiting for local unions page to load: {e}")
-            raise e
+            # Wait briefly for the PDF list to load
+            await page.wait_for_selector('a[href$=".pdf"]', timeout=5000)
+        except Exception:
+            # No PDF links on the page
+            return pdf_links
 
-        soup: BeautifulSoup = BeautifulSoup(await page.content(), "html.parser")
-        union_details: List[UnionDetail] = []
-        accordion_items: Sequence[Tag] = cast(
-            Sequence[Tag],
-            soup.find_all(
-                "div", attrs={"class": "wp-block-ns-accordion__item"})
-        )
-        for item in accordion_items:
-            button_elem = cast(Optional[Tag], item.find(
-                "button", attrs={"class": "wp-block-ns-accordion__item-toggle"}))
-            campus: str = button_elem.get_text(
-                strip=True) if button_elem else "Unknown"
-            links: Sequence[Tag] = cast(Sequence[Tag], item.find_all("a"))
-            for a_tag in links:
-                href = a_tag.get("href")
-                link_text: str = cast(str, a_tag.get_text(strip=True))
-                if href and isinstance(href, str) and "—" in link_text:
-                    name, code = link_text.split("—")
-                    union_detail = UnionDetail(
-                        name=name.strip(),
-                        code=code.strip(),
-                        url=href,
-                        scope=campus
-                    )
-                    union_details.append(union_detail)
-        return union_details
+        # Query all anchors with .pdf in href
+        anchors = await page.query_selector_all('a[href$=".pdf"]')
+        seen = set()
+        for a in anchors:
+            href = await a.get_attribute("href")
+            if not href:
+                continue
+            # Normalize relative URLs to absolute
+            if href.startswith("/"):
+                href = f"https://ucnet.universityofcalifornia.edu{href}"
+            # Filter duplicates
+            if href not in seen:
+                seen.add(href)
+                pdf_links.append(href)
+        return pdf_links
 
-    async def get_systemwide_unions(self, url: str, page: Page) -> List[UnionDetail]:
-        """Retrieve systemwide union details from the provided URL."""
-        await page.goto(url)
-        try:
-            await page.wait_for_selector("#systemwide", timeout=10000)
-        except PlaywrightTimeoutError as e:
-            logger.error(
-                f"Error waiting for systemwide unions page to load: {e}")
-            raise e
-
-        soup: BeautifulSoup = BeautifulSoup(await page.content(), "html.parser")
-        union_details: List[UnionDetail] = []
-        a_tags: Sequence[Tag] = cast(
-            Sequence[Tag],
-            soup.find_all("a", attrs={"href": True})
-        )
-        for a_tag in a_tags:
-            href = a_tag.get("href")
-            scope: str = "ucop"
-            span_tags: Sequence[Tag] = cast(
-                Sequence[Tag], a_tag.find_all("span"))
-            full_text: str = ""
-            if span_tags:
-                full_text = cast(str, span_tags[0].get_text(strip=True))
-            if href and isinstance(href, str) and "—" in full_text:
-                name, code = full_text.split("—")
-                union_detail = UnionDetail(
-                    name=name.strip(),
-                    code=code.strip(),
-                    url=href,
-                    scope=scope
-                )
-                union_details.append(union_detail)
-        return union_details
-
-    async def get_local_union_contracts(self, unions: List[UnionDetail], page: Page) -> List[DocumentDetails]:
-        """
-        For each local union, navigate to its URL and extract contract links as PolicyDetails.
-        """
+    async def get_systemwide_units(self, page: Page) -> List[DocumentDetails]:
+        """Extract systemwide bargaining unit contracts."""
         policy_details_list: List[DocumentDetails] = []
-        for union in unions:
-            await page.goto(union.url)
-            try:
-                await page.wait_for_selector("#content-detail__content", timeout=10000)
-            except PlaywrightTimeoutError as e:
-                logger.error(
-                    f"Error waiting for contract detail page for union {union.name}: {e}")
+
+        # Navigate to the landing page
+        await page.goto(BARGAINING_UNITS_URL)
+        await page.wait_for_load_state("networkidle")
+
+        logger.info(f"Navigated to {BARGAINING_UNITS_URL}")
+
+        # The systemwide units are in accordion sections under the "Systemwide bargaining units by union" heading
+        accordion_sections = await page.query_selector_all(".wp-block-ns-accordion__item")
+        logger.info(f"Found {len(accordion_sections)} accordion sections")
+
+        for i, accordion in enumerate(accordion_sections):
+            button = await accordion.query_selector(".wp-block-ns-accordion__item-toggle")
+            if not button:
                 continue
 
-            soup: BeautifulSoup = BeautifulSoup(await page.content(), "html.parser")
-            content_detail = cast(Optional[Tag], soup.find(
-                id="content-detail__content"))
-            if content_detail:
-                a_tags: Sequence[Tag] = cast(
-                    Sequence[Tag],
-                    content_detail.find_all("a", attrs={"href": True})
-                )
-                for a_tag in a_tags:
-                    href = cast(Optional[str], a_tag.get("href"))
-                    if href and isinstance(href, str) and href.endswith(".pdf"):
-                        title: str = href.split("/")[-1].replace(".pdf", "")
+            union_name = (await button.inner_text()).strip()
+            logger.info(f"Processing accordion {i}: {union_name}")
+
+            # Determine if this is a systemwide union or a local campus
+            # Campus names typically contain "UC" followed by campus name
+            # not the best method but works for current structure
+            is_campus = any(campus_indicator in union_name for campus_indicator in [
+                "UC Berkeley", "UC Davis", "UC Irvine", "UC Los Angeles", "UCLA",
+                "UC Merced", "UC Riverside", "UC San Diego", "UC Santa Barbara",
+                "UC Santa Cruz", "UC San Francisco", "Lawrence Berkeley"
+            ])
+
+            if is_campus:
+                # This is a local campus accordion, skip it in systemwide processing
+                logger.info(
+                    f"Skipping campus accordion in systemwide: {union_name}")
+                continue
+
+            # Click to expand the accordion if it's not already expanded
+            try:
+                await button.click()
+                await page.wait_for_timeout(500)  # Wait for expansion
+            except:
+                pass
+
+            # Look for links in this accordion section
+            links = await accordion.query_selector_all("a")
+            logger.info(f"Found {len(links)} links in {union_name}")
+
+            for link in links:
+                link_text = (await link.inner_text()).strip()
+                href = await link.get_attribute("href")
+
+                if not href or not link_text:
+                    continue
+
+                logger.info(f"Processing systemwide link: {link_text}")
+
+                # Parse the link text to extract code and name (format: "CODE — Name")
+                match = re.match(r"^(\w+)\s+—\s+(.*)", link_text)
+                if not match:
+                    logger.warning(
+                        f"Link text doesn't match expected format: {link_text}")
+                    continue
+
+                code, name = match.groups()
+                logger.info(f"Found systemwide unit: {code} - {name}")
+
+                # The href should be the contract page URL
+                contract_url = href
+                if not contract_url.startswith("http"):
+                    contract_url = f"https://ucnet.universityofcalifornia.edu{contract_url}"
+
+                # Open the contract page in a new tab
+                contract_page = await page.context.new_page()
+                try:
+                    await contract_page.goto(contract_url, timeout=30000)
+                    await contract_page.wait_for_load_state("networkidle")
+                    pdf_links = await self.extract_pdf_links(contract_page)
+
+                    logger.info(f"Found {len(pdf_links)} PDFs for {code}")
+
+                    # Create DocumentDetails for each PDF
+                    for pdf_url in pdf_links:
+                        title = pdf_url.split("/")[-1].replace(".pdf", "")
+                        logger.info(
+                            f"Creating DocumentDetails for: {title} from {pdf_url}")
                         policy_detail = DocumentDetails(
                             title=title,
-                            url=href,
+                            url=pdf_url,
                             metadata={
-                                "keywords": [union.code, union.name, union.scope],
-                                "subject_areas": ["Collective Bargaining", union.code],
-                                "responsible_office": union.scope
+                                "keywords": [code, name, "ucop"],
+                                "subject_areas": ["Collective Bargaining", code],
+                                "responsible_office": "ucop"
                             }
                         )
                         policy_details_list.append(policy_detail)
+                except Exception as e:
+                    logger.error(
+                        f"Error processing systemwide unit {code}: {e}")
+                finally:
+                    await contract_page.close()
+
+        logger.info(
+            f"Total systemwide policy details: {len(policy_details_list)}")
         return policy_details_list
 
-    async def get_systemwide_union_contracts(self, unions: List[UnionDetail], page: Page) -> List[DocumentDetails]:
-        """
-        For each systemwide union, navigate to its contract page and extract contract links as PolicyDetails.
-        """
+    async def get_local_units(self, page: Page) -> List[DocumentDetails]:
+        """Extract local bargaining unit contracts."""
         policy_details_list: List[DocumentDetails] = []
-        for union in unions:
-            contract_url: str = f"{union.url}/contract/"
-            await page.goto(contract_url)
-            try:
-                await page.wait_for_selector("#content-detail__content", timeout=10000)
-            except PlaywrightTimeoutError as e:
-                logger.error(
-                    f"Error waiting for contract detail page for union {union.name}: {e}")
+
+        # Extract local units - they are in the accordion sections that represent campuses
+        accordion_sections = await page.query_selector_all(".wp-block-ns-accordion__item")
+        logger.info(
+            f"Processing {len(accordion_sections)} accordion sections for local units")
+
+        for i, accordion in enumerate(accordion_sections):
+            button = await accordion.query_selector(".wp-block-ns-accordion__item-toggle")
+            if not button:
                 continue
 
-            soup: BeautifulSoup = BeautifulSoup(await page.content(), "html.parser")
-            content_detail = cast(Optional[Tag], soup.find(
-                id="content-detail__content"))
-            if content_detail:
-                a_tags: Sequence[Tag] = cast(
-                    Sequence[Tag],
-                    content_detail.find_all("a", attrs={"href": True})
-                )
-                for a_tag in a_tags:
-                    href = cast(Optional[str], a_tag.get("href"))
-                    if href and isinstance(href, str) and href.endswith(".pdf"):
-                        title: str = href.split("/")[-1].replace(".pdf", "")
+            union_name = (await button.inner_text()).strip()
+            logger.info(
+                f"Checking accordion {i} for local units: {union_name}")
+
+            # Determine if this is a local campus accordion
+            # Campus names typically contain "UC" followed by campus name
+            campus_indicators = [
+                "UC Berkeley", "UC Davis", "UC Irvine", "UC Los Angeles", "UCLA",
+                "UC Merced", "UC Riverside", "UC San Diego", "UC Santa Barbara",
+                "UC Santa Cruz", "UC San Francisco", "Lawrence Berkeley"
+            ]
+
+            is_campus = any(
+                campus_indicator in union_name for campus_indicator in campus_indicators)
+
+            if not is_campus:
+                # This is not a campus accordion, skip it in local processing
+                continue
+
+            campus_name = union_name
+            logger.info(f"Processing local campus: {campus_name}")
+
+            # Click to expand the accordion if it's not already expanded
+            try:
+                await button.click()
+                await page.wait_for_timeout(500)  # Wait for expansion
+            except:
+                pass
+
+            # Look for links in this accordion section
+            links = await accordion.query_selector_all("a")
+            logger.info(f"Found {len(links)} links in {campus_name}")
+
+            for link in links:
+                link_text = (await link.inner_text()).strip()
+                href = await link.get_attribute("href")
+
+                if not href or not link_text:
+                    continue
+
+                logger.info(f"Processing local link: {link_text}")
+
+                # Parse the link text to extract code and name (format: "CODE — Name")
+                match = re.match(r"^(\w+)\s+—\s+(.*)", link_text)
+                if not match:
+                    logger.warning(
+                        f"Link text doesn't match expected format: {link_text}")
+                    continue
+
+                code, name = match.groups()
+                logger.info(
+                    f"Found local unit: {code} - {name} at {campus_name}")
+
+                # The href should be the contract page URL
+                contract_url = href
+                if not contract_url.startswith("http"):
+                    contract_url = f"https://ucnet.universityofcalifornia.edu{contract_url}"
+
+                # Open the contract page in a new tab
+                contract_page = await page.context.new_page()
+                try:
+                    await contract_page.goto(contract_url, timeout=30000)
+                    await contract_page.wait_for_load_state("networkidle")
+                    pdf_links = await self.extract_pdf_links(contract_page)
+
+                    logger.info(
+                        f"Found {len(pdf_links)} PDFs for {code} at {campus_name}")
+
+                    # Create DocumentDetails for each PDF
+                    for pdf_url in pdf_links:
+                        title = pdf_url.split("/")[-1].replace(".pdf", "")
                         policy_detail = DocumentDetails(
                             title=title,
-                            url=href,
+                            url=pdf_url,
                             metadata={
-                                "keywords": [union.code, union.name, union.scope],
-                                "subject_areas": ["Collective Bargaining", union.code],
-                                "responsible_office": union.scope
+                                "keywords": [code, name, campus_name],
+                                "subject_areas": ["Collective Bargaining", code],
+                                "responsible_office": campus_name
                             }
                         )
                         policy_details_list.append(policy_detail)
+                except Exception as e:
+                    logger.error(
+                        f"Error processing local unit {code} at {campus_name}: {e}")
+                finally:
+                    await contract_page.close()
+
+        logger.info(f"Total local policy details: {len(policy_details_list)}")
         return policy_details_list
 
     async def get_uc_collective_bargaining_links(self, page: Page) -> List[DocumentDetails]:
         """
-        Combines local and systemwide union contracts into a single list of PolicyDetails.
+        Combines systemwide and local union contracts into a single list of DocumentDetails.
         """
-        home_url: str = base_url
-        local_unions: List[UnionDetail] = await self.get_local_unions(home_url, page)
-        systemwide_unions: List[UnionDetail] = await self.get_systemwide_unions(home_url, page)
         policy_details: List[DocumentDetails] = []
-        policy_details += await self.get_local_union_contracts(local_unions, page)
-        policy_details += await self.get_systemwide_union_contracts(systemwide_unions, page)
+        policy_details += await self.get_systemwide_units(page)
+        policy_details += await self.get_local_units(page)
         return policy_details
 
     async def __aiter__(self) -> AsyncIterator[DocumentDetails]:
         """
         Asynchronously iterate over the collective bargaining contracts.
         Launches the browser, navigates to the UCOP Collective Bargaining page,
-        retrieves union details and contracts, and yields each contract as a PolicyDetails object.
+        retrieves union details and contracts, and yields each contract as a DocumentDetails object.
         """
         async with get_browser_page(user_agent) as page:
             policy_links: List[DocumentDetails] = await self.get_uc_collective_bargaining_links(page)
@@ -211,15 +282,23 @@ class UcopCollectiveBargainingStream(DocumentStream):
 
 
 if __name__ == "__main__":
+    print("Starting cb.py script...")
     source = Source(
         name="UCCONTRACTS",
         url="https://ucnet.universityofcalifornia.edu/resources/employment-policies-contracts/bargaining-units/",
         last_updated=None,
-        type="UCCONTRACTS")
+        type="UCCONTRACTS"
+    )
 
     async def main():
-        async for doc in UcopCollectiveBargainingStream(source):
-            print(doc)
-            print(doc.metadata)
+        print("Starting async main...")
+        count = 0
+        async for doc in UcnetCollectiveBargainingStream(source):
+            count += 1
+            print(f"Document {count}: {doc.title}")
+            print(f"URL: {doc.url}")
+            print(f"Metadata: {doc.metadata}")
+            print("---")
+        print(f"Total documents found: {count}")
 
     asyncio.run(main())
